fix:기수 활성화 상태 변경 플로우 점검
