+++ src/queries/operation/useEditGenerationMutation.ts
@@ -1,8 +1,7 @@
 import { useMutation } from '@tanstack/react-query';
 import { AxiosError, AxiosResponse } from 'axios';
-import { ErrorResponse } from 'react-router-dom';
 
-import { ApiResponse } from 'apis/common/types';
+import { ApiResponse, ErrorResponse } from 'apis/common/types';
 import { patchGenerationActive } from 'apis/operation/OperationApis';
 import { EditGenerationReq, EditGenerationRes } from 'apis/operation/types';
 
@@ -17,7 +16,19 @@ export const useEditGenerationMutation = () => {
       console.log('res :', res);
     },
     onError: (err) => {
-      console.log('err :', err);
+      if (err.response) {
+        const errorData = err.response.data;
+        if (err.response.status === 404) {
+          if (errorData.errorCode === 'OPR_1000') {
+            window.alert(`${errorData.errorCode}\n${errorData.message}`);
+          }
+        } else if (err.response.status === 409) {
+          if (errorData.errorCode === 'OPR_1001') {
+            window.alert(`${errorData.errorCode}\n${errorData.message}`);
+          }
+        }
+        console.error('login error ', errorData.message);
+      }
     },
   });
 };
